refactor: extract multi-step application form shell
